requirements.txt
attrs==18.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-extensions==2.0.7
django-money==0.13.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
drf-nested-routers==0.90.2
factory-boy==2.11.1
Faker==0.8.15
inflection==0.3.1
itypes==1.1.0
more-itertools==4.1.0
openapi-codec==1.3.2
pluggy==0.6.0
py==1.5.3
py-moneyed==0.7.0
pytest==3.5.1
pytest-factoryboy==2.0.1
simplejson==3.15.0
text-unidecode==1.2
uritemplate==3.0.0
Werkzeug==0.14.1
